scripts: some adjustments to the git hooks
Qu pointed out that codespell complains about 'uptodate' despite us
using it everywhere, so add it to an ignore list and add that to the
command.

Also the checkpatch thing is a little heavy handed, it blew up on a lot
of his patches where it wasn't really necessary or applied to code he
was actually changing.  Fix the hook to allow the user to bypass the
check if they don't think checkpatch is correct.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>