create-issue: retry if user doesn't have project permissions
For btrfs patch submitters who do not yet have project permissions,
btrfs-create-issue currently fails altogether due to lack of permission.
This makes their patch mailing-list-only, making it harder to work in
the project workflow. Adding a retry without the project set means an
issue will be created; the issue will be outside the project, but at
least it exists, making it easier to add to the project workflow.