workflow: btrfs-send-patches: ignore cases for 'Message-ID'
[BUG]
With git updated to 2.41.0, btrfs-send-patches no longer works:

$ ~/btrfs/workflow/scripts/btrfs-send-patches --to linux-btrfs@vger.kernel.org 0001-btrfs-do-not-ASSERT-on-duplicated-global-roots.patch
Message-Id wasn't present in the patch provided

[CAUSE]
Git 2.41.0 changed the tag 'Message-Id' to 'Message-ID' (note the
tailing 'd' is changed to upper case).

But the script btrfs-send-patches is checking the message-id using the
old format, thus it failed to locate the message id of the newly
generated patches.

[FIX]
Update btrfs-send-patches scripts to ignore cases on 'Message-ID', this
involves two parts:

- Grep of 'Message-ID'
- Python regex matching 'Message-ID'

Signed-off-by: Qu Wenruo <wqu@suse.com>